manim-animator/requirements.txt
fastapi
uvicorn
python-dotenv
langchain-openai
langgraph
manim
